parsing work - still buggy
