+++ AndroidManifest.xml
@@ -95,7 +95,7 @@
 
          <activity
             android:name="org.cirdles.chroni.TablePainter"
-            android:parentActivityName="android.intent.action.DISPLAY"
+            android:parentActivityName="android.intent.action.MAINMENU"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.DISPLAY" />
