+++ res/layout-large/aliquot_select.xml
@@ -98,7 +98,7 @@
          android:layout_height="wrap_content"
          android:layout_alignBottom="@+id/aliquotFileSelectText"
          android:layout_alignRight="@+id/aliquotIGSNText"
-         android:background="@color/grey"
+         android:background="@color/dark_grey"
          android:minHeight="48dp"
          android:minWidth="48dp"
          android:text="+"
