+++ res/layout/aliquot_select.xml
@@ -34,7 +34,7 @@
         android:layout_height="wrap_content"
         android:layout_alignRight="@+id/aliquotIGSNText"
         android:layout_alignTop="@+id/aliquotFileSubmitButton"
-        android:background="@color/grey"
+        android:background="@color/dark_grey"
         android:minHeight="48dp"
         android:minWidth="48dp"
         android:text="+"
