+++ res/layout/report_settings_select.xml
@@ -80,7 +80,7 @@
         android:layout_alignBottom="@+id/textView2"
         android:layout_marginLeft="51dp"
         android:layout_toRightOf="@+id/imageView1"
-        android:background="@color/grey"
+        android:background="@color/dark_grey"
         android:minHeight="48dp"
         android:minWidth="48dp"
         android:text="+"
